+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/bean/key/FastJsonDenseUuidKey.java
@@ -1,8 +1,8 @@
 package com.dwarfeng.subgrade.sdk.bean.key;
 
 import com.alibaba.fastjson.annotation.JSONField;
-import com.dwarfeng.subgrade.stack.bean.Bean;
 import com.dwarfeng.subgrade.stack.bean.key.DenseUuidKey;
+import com.dwarfeng.subgrade.stack.bean.key.Key;
 import org.springframework.lang.NonNull;
 
 import java.util.Objects;
@@ -13,7 +13,7 @@ import java.util.Objects;
  * @author DwArFeng
  * @since 0.0.1-beta
  */
-public class FastJsonDenseUuidKey implements Bean {
+public class FastJsonDenseUuidKey implements Key {
 
     private static final long serialVersionUID = -678760030501902949L;